+++ src/doomstat.c
@@ -168,6 +168,7 @@ int a11y_invul_colormap;
 
 // Weapons
 int no_hor_autoaim;
+int switch_on_pickup;
 int weapon_bobbing_percentage;
 int bobbing_style;
 int weapon_inertia;